A centre position for servos added
When used in anger, it turned out to be useful to start each servo in a central position and adjust from there. A constant can be set to have servos start in the centre - not a good idea if trains are running! You also have a menu option to centre them all.